Time: 7 ms (18.74%) | Memory: 7.8 MB (21.72%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cpp
@@ -6,8 +6,8 @@ bool isperfectsquare(int n){
     else return false;
 }
     bool judgeSquareSum(int c) {
-        long long x = 0;
-        long long y = c;
+        int x = 0;
+        int y = c;
         
         while(x<=y){
             if(isperfectsquare(x) && isperfectsquare(y)) return true;
